split parser in two files
